+++ src/components/NavBar/NavBarView.js
@@ -4,7 +4,7 @@ import {
   navBarData, navBarCartData, navBarstyling,
 } from '../../bento/navigationBarData';
 import Login from '../GoogleAuth/loginComponent';
-import globalData from '../../bento/globalData';
+import globalData from '../../bento/siteWideConfig';
 
 const BentoNavBar = ({ cartFieldIds }) => (
   <>
