+++ src/pages/armDetail/armDetailView.js
@@ -11,7 +11,7 @@ import StatsView from '../../components/Stats/StatsView';
 import { Typography } from '../../components/Wrappers/Wrappers';
 import icon from '../../assets/icons/Arms.Icon.svg';
 import fileCountIcon from '../../assets/icons/Program_Detail.FileCount.svg';
-import globalData from '../../bento/globalData';
+import globalData from '../../bento/siteWideConfig';
 import {
   header,
   subsections,
